source/assign4_stepanovic/requirement.txt
Django version 1.11.5
Python 3.6.2
PyJWT 1.5.3
certifi 2017.7.27.1
chardet 3.0.4
django-common-helpers 0.9.1
django-cron 0.5.0
django-discover-runner 1.0
django-extensions 1.9.6
django-sslserver 0.20
djangorestframework 3.7.0
djangorestframework-jwt  1.11.0
idna 2.6
pip 9.0.1
pytz 2017.2
requests 2.18.4
rest-framework-auth0 0.4.5
setuptools 28.8.0
six 1.11.0
urllib3 1.22
